requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
blessings==1.7
click==7.1.2
curtsies==0.3.4
dj-rest-auth==5.0.2
Django==4.1
django-cli-g73==0.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
enquiries==0.1.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyfiglet==0.8.post1
PyMySQL==1.1.0
python-decouple==3.8
pytz==2023.3.post1
six==1.15.0
sqlparse==0.4.4
typing_extensions==4.8.0
wcwidth==0.2.5
whitenoise==6.6.0
